PixelInformation window works fine. All what is left is to set a proper startup location for it.

+++ ImageViewer/ImageViewer/Model/Event/SendPixelInformationEvent.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 
 namespace ImageViewer.Model.Event
 {
-    class SendPixelInformationEvent : PubSubEvent<Dictionary<string, byte>>
+    class SendPixelInformationEvent : PubSubEvent<Dictionary<string, int>>
     {
 
     }
